move these tests in test_callop.py
